requirements.txt
azure-functions
azure-cosmos
azure-communication-sms
requests
circle-sdks